Implement ScenarioGenerator
The scenario generator is the piece that is responsible
for painting a scenario that the arranger can work with.
It generates a data set of all drivers and passengers
based on source and destination locations. The source and
destination locations are determined by the scenario rules
defined by the user.
NOTE: There are still a few more things to work on before
      the arranger logic can be worked on.

TODO:
    * How will the arranger decide to form the groups? Will
      it do it based on dest or src location?
    * How do we handle different destination locations? For
      example, if a group of people are leaving from the same
      location, but a subset of the people are leaving for a
      different location.
    * How to choose drivers - random or based on a schedule?